+++ Changelog.md
@@ -2,7 +2,8 @@
 
 ## V 1.2.0 : XXX Update - XX/XX/19
 
-    - CameraSystem : Create
+    - All : Use property decorator
+    - CameraSystem : Create basic camera
     - GameState : Rename to World
     - Window : Modify management of Worlds